third_party/blink/web_tests/external/wpt/webrtc/RTCRtpEncodingParameters-codec-opus-stereo.https-expected.txt

This is a testharness.js-based test.
[FAIL] Sending and receiving a mono track with opus (stereo decoder upmix)
  assert_equals: Remote track is stereo (despite local track being mono) expected 2 but got 1
[FAIL] Sending and receiving a stereo track with opus
  assert_equals: Remote track is also stereo expected 2 but got 1
Harness: the test ran to completion.
